2019_Presentation
During the fall semester of my senior I was assigned a semester-long consulting group project in my International Finance course.  The objective of the project was to analyze the operations of a multinational firm that was assigned to us in the following countries/regions: China, Europe, Mexico, and South America.  We were to develop strategies to minimize risk and then present to the executives and student peers with the profit-taking opportunities that might be available.  

My group was assigned a firm called Federated Investors.  In particular, I was the team captain and had the responsibilities of communicating with the executive of Federated Investors to schedule a review meeting of our ideas before our final presentation, delegating tasks to group members, and ensuring that our team was on-track to completing the project.