Report_2019
The objective of this project was to develop a concise and comprehensive strategic plan for Sears Holding Corporation for my Business Strategy and Policy course. My partner and I analyzed the mission and vision statements as well as the historic and present financial standing of Sears, evaluated the company’s resources and competitive position, completed an internal and external factor evaluation matrix, analyzed the macro and industry environments, created a SWOT analysis, stated our strategic objectives and strategic formulation, and proposed an implementation plan and evaluation.  In addition, a formal presentation was given for peers and the professor.