Report_2015
For my English research paper in my junior year of high school, I wrote about the economic effects and impacts of WW2. 

Thesis: World War II positively affected the United States economy because the war brought an end to the Great Depression, initiated an industry of innovations and prosperity, and stimulated the beginning of American imperialism.