Report_2013
In my American History class during my freshman year of high school, I was asked to watch the 1939 Wizard of Oz movie and analyze how the movie is an allegory.  Specifically, I analyzed how the symbols, characters, and scenes in the movie representative with the economic and political problems of the Populist Movement and the Election of 1896.