Report_2016
During my freshmen year of college, I wrote a paper for my U.S. History class about the political impacts of the American Civil War and Reconstruction period. 

Thesis: The era during the American Civil War profoundly altered the United States and turned the Reconstruction age into a political revolution for a united American nation because of new political ideas of the southern Republican controlled congress, the new amendments of the United States Constitution, and the centralization of government.