Report_2019
For my Business Ethics class, I wrote a case study analyzing the application of Catholic Social Teaching and the Benedictine Hallmarks on the Facebook-Cambridge Analytica Scandal. 

Thesis: To analyze the Facebook-Cambridge Analytica scandal, the core values of Catholic Social Teaching (Human Dignity, the Common Good, Subsidiarity, Solidarity, and the Universals Destination of Goods) and four of the ten Benedictine hallmarks (Stability, Community, Hospitality, and Stewardship) will be applied to the case.